+++ app/src/main/java/com/romanpulov/odeon/db/Composition.java
@@ -12,19 +12,19 @@ public class Composition {
     @PrimaryKey
     @NonNull
     @ColumnInfo(name = "comp_id")
-    private Integer id;
+    private Long id;
 
     @NonNull
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
     @NonNull
     @ColumnInfo(name = "artf_id")
-    private Integer artifactId;
+    private Long artifactId;
 
     @NonNull
-    public Integer getArtifactId() {
+    public Long getArtifactId() {
         return artifactId;
     }
 
@@ -64,7 +64,7 @@ public class Composition {
         return number;
     }
 
-    public Composition(@NonNull Integer id, @NonNull Integer artifactId, @NonNull String title, @Nullable Integer duration, @Nullable Integer diskNumber, @Nullable Integer number) {
+    public Composition(@NonNull Long id, @NonNull Long artifactId, @NonNull String title, @Nullable Integer duration, @Nullable Integer diskNumber, @Nullable Integer number) {
         this.id = id;
         this.artifactId = artifactId;
         this.title = title;
